+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -1,5 +1,9 @@
 ## v2.43.0-dev
 
+* Added support for generator functions to `ControlFlow#execute` and
+    `ControlFlow#wait`. For more information, see documentation on
+    `webdriver.promise.consume`. Requires harmony support (run with
+    `node --harmony-generators` in `v0.11.x`).
 * Promise A+ compliance: a promise may no longer resolve to itself.
 * For consistency with other language bindings, deprecated
     `UnhandledAlertError#getAlert` and added `#getAlertText`.